+++ src/main/java/racingcar/dao/PlayerDao.java
@@ -1,54 +1,11 @@
 package racingcar.dao;
 
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
 import java.util.List;
-import org.springframework.jdbc.core.BatchPreparedStatementSetter;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.stereotype.Repository;
 import racingcar.entity.Player;
 
-@Repository
-public class PlayerDao {
+public interface PlayerDao {
 
-    private final JdbcTemplate jdbcTemplate;
-    private final RowMapper<Player> playerRowMapper = (resultSet, rowNum) ->
-            new Player(
-                    resultSet.getInt("id"),
-                    resultSet.getString("name"),
-                    resultSet.getInt("position"),
-                    resultSet.getInt("game_id")
-            );
+    void insert(List<Player> players);
 
-    public PlayerDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public void insert(List<Player> players) {
-        
-        String sql = "insert into PLAYER (name, position, game_id) values (?, ?, ?)";
-
-        jdbcTemplate.batchUpdate(sql, new BatchPreparedStatementSetter() {
-
-            @Override
-            public void setValues(PreparedStatement ps, int i) throws SQLException {
-                Player player = players.get(i);
-                ps.setString(1, player.getName());
-                ps.setInt(2, player.getPosition());
-                ps.setInt(3, player.getGameId());
-            }
-
-            @Override
-            public int getBatchSize() {
-                return players.size();
-            }
-        });
-    }
-
-    public List<Player> findAll() {
-        String sql = "select * from PLAYER";
-
-        return jdbcTemplate.query(sql, playerRowMapper);
-    }
+    List<Player> findAll();
 }
